+++ mod.ts
@@ -1,27 +1,54 @@
-export { default as Button } from './src/atoms/Button/index.tsx';
-export { default as Card } from './components/Card/index.tsx';
-export { default as Chip } from './components/Chip/index.tsx';
-export { default as Chiplist } from './components/Chiplist/index.tsx';
-export { default as Code } from './src/atoms/Code/index.tsx';
-export { default as Dialog } from './components/Dialog/index.tsx';
-export { default as Fieldset } from './components/Fieldset/index.tsx';
-export { default as Footer } from './components/Footer/index.tsx';
-export { default as Gradient } from './components/Gradient/index.tsx';
-export { default as Header } from './components/Header/index.tsx';
-export { default as Input } from './components/Input/index.tsx';
-export { default as Kbd } from './src/atoms/Kbd/index.tsx';
-export { default as Layout } from './src/atoms/Layout/index.tsx';
-export { default as Link } from './src/atoms/Link/index.tsx';
-export { default as Linkmap } from './components/Linkmap/index.tsx';
-export { default as Loader } from './components/Loader/index.tsx';
-export { default as Main } from './components/Main/index.tsx';
-export { default as Markdown } from './src/atoms/Markdown/index.tsx';
-export { default as Menu } from './components/Menu/index.tsx';
-export { default as Navigation } from './components/Navigation/index.tsx';
-export { default as Panel } from './src/atoms/Panel/index.tsx';
-export { default as Select } from './components/Select/index.tsx';
-export { default as Separator } from './src/atoms/Separator/index.tsx';
+//   __  __         _      _
+//  |  \/  |___  __| |_  _| |___
+//  | |\/| / _ \/ _` | || | / -_)
+//  |_|  |_\___/\__,_|\_,_|_\___|
+//
+////////////////////////////////////////////////////////////////////////////////////////////////////////
+/**
+ * Inside here all atom and molecule components are exported. Because organisms are, by nature, islands
+ * they must not be exported here. Instead, organisms must be forwareded to the `islands/` directory
+ * using the Lunchbox plugin. Components exported here must be SSR and non interactible.
+ *
+ * @module
+ */
+// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+//     _  _
+//    /_\| |_ ___ _ __  ___
+//   / _ \  _/ _ \ '  \(_-<
+//  /_/ \_\__\___/_|_|_/__/
+//
+// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+export { default as Button } from './src/components/atoms/Button/index.tsx';
+export { default as Code } from './src/components/atoms/Code/index.tsx';
+export { default as Kbd } from './src/components/atoms/Kbd/index.tsx';
+export { default as Layout } from './src/components/atoms/Layout/index.tsx';
+export { default as Link } from './src/components/atoms/Link/index.tsx';
+export { default as Markdown } from './src/components/atoms/Markdown/index.tsx';
+export { default as Panel } from './src/components/atoms/Panel/index.tsx';
+export { default as Separator } from './src/components/atoms/Separator/index.tsx';
+export { default as Text } from './src/components/atoms/Text/index.tsx';
+// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+//   __  __     _             _
+//  |  \/  |___| |___ __ _  _| |___ ___
+//  | |\/| / _ \ / -_) _| || | / -_|_-<
+//  |_|  |_\___/_\___\__|\_,_|_\___/__/
+//
+// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+export { default as Card } from './src/components/molecules/Card/index.tsx';
+export { default as Chip } from './src/components/molecules/Chip/index.tsx';
+export { default as Fieldset } from './src/components/molecules/Fieldset/index.tsx';
+export { default as Footer } from './src/components/molecules/Footer/index.tsx';
+export { default as Header } from './src/components/molecules/Header/index.tsx';
+export { default as Input } from './src/components/molecules/Input/index.tsx';
+export { default as Main } from './src/components/molecules/Main/index.tsx';
+export { default as Navigation } from './src/components/molecules/Navigation/index.tsx';
+export { default as Select } from './src/components/molecules/Select/index.tsx';
+export { default as TextArea } from './src/components/molecules/TextArea/index.tsx';
+// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+//   ___                           _          _
+//  |   \ ___ _ __ _ _ ___ __ __ _| |_ ___ __| |
+//  | |) / -_) '_ \ '_/ -_) _/ _` |  _/ -_) _` |
+//  |___/\___| .__/_| \___\__\__,_|\__\___\__,_|
+//           |_|
+// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 export { default as Stylesheet } from './components/Stylesheet/index.tsx';
-export { default as Text } from './src/atoms/Text/index.tsx';
-export { default as TextArea } from './components/TextArea/index.tsx';
-export { default as Time } from './components/Time/index.tsx';