+++ src/components/molecules/Input/index.tsx
@@ -1,0 +1,91 @@
+//   ___                _
+//  |_ _|_ _  _ __ _  _| |_
+//   | || ' \| '_ \ || |  _|
+//  |___|_||_| .__/\_,_|\__|
+//           |_|
+////////////////////////////////////////////////////////////////////////////////////////////////////////
+/**
+ * ### Input
+ * *Molecule*
+ *
+ * This module contains the render function for the `<Input />` component.
+ *
+ * @module
+ */
+import Text from '../../atoms/Text/index.tsx';
+import setup, { iInput } from './setup.ts';
+
+// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+/**
+ * Render function for the [`<Input/ >`](/x/lunchbox/components/Input/setup.ts?s=iInput) component.
+ *
+ * @param {Partial<iInput>} props
+ *  {@link iInput} (Partial by [design](https://deno.land/x/lunchbox#configure-anything-easily))
+ *
+ * @returns {JSXInternal.Element}
+ *  The `<Input />` component.
+ */
+export default function (props: Partial<iInput>) {
+  const {
+    c,
+    nostyle,
+    nostyleAll,
+    fref,
+    fwd,
+    maxWidth,
+    label,
+    fieldIcon,
+    error,
+    ...p
+  } = setup(props);
+
+  return (
+    <div ref={fwd.container?.ref} class={c.container}>
+      <label ref={fwd.label?.ref} class={c.label}>
+        {label === '' ? null : (
+          <Text
+            nostyleAll={nostyleAll}
+            fref={fwd.text?.fref}
+            noMargins
+            class={c.text}
+          >
+            <>{label}</>
+            {p.required
+              ? (
+                <sup
+                  ref={fwd.required?.ref}
+                  title='Required'
+                  class={c.required}
+                >
+                  *
+                </sup>
+              )
+              : null}
+          </Text>
+        )}
+        {fieldIcon
+          ? (
+            <div class={c.iconContainer}>
+              {fieldIcon}
+            </div>
+          )
+          : null}
+        <input ref={fref} class={c.input} {...p} />
+      </label>
+      {error
+        ? (
+          <Text
+            fref={fwd.error?.fref}
+            noMargins
+            type='small'
+            inheritColor
+            class={c.error}
+            nostyleAll={nostyleAll}
+          >
+            {error}
+          </Text>
+        )
+        : null}
+    </div>
+  );
+}