+++ src/components/organisms/Chiplist/index.tsx
@@ -1,0 +1,26 @@
+import setup, { iChiplist } from './setup.ts';
+import Chip from '../../molecules/Chip/index.tsx';
+
+export default function (props: Partial<iChiplist>) {
+  const { c, nostyle, nostyleAll, fref, fwd, values, onRemove, ...p } = setup(
+    props,
+  );
+
+  return (
+    <ul ref={fref} class={c.list} {...p}>
+      {values.map((value) => (
+        <Chip
+          class={c.chip}
+          onRemove={onRemove}
+          fwd={{
+            remove_button: {
+              class: c.remove_button,
+            },
+          }}
+        >
+          {value}
+        </Chip>
+      ))}
+    </ul>
+  );
+}